Create README.md
A News Aggregator for Literary Publications

Download and run on XCode

Tapestry finds and brings together the best short fiction, poetry and essays and book reviews on the web. Think of it as a "Google News" for book-lovers and writers. Gathering work from over 100 sources and counting, the app is designed to help readers fall in love with up and coming authors, and it allows publishers and authors to grow their websites by making their work easy to find and easy to use. If you know anybody who likes a good book, then they will enjoy this a lot.

I created this app because (1) it's extremely hard for independent authors and publications to grow online and (2) the literary publishing industry is a billion-dollar market lacking a viable, meritocratic platform. Tapestry will let them connect and submit work to all these publications. Now, you can submit fiction to over 60 literary magazines in seconds.

But I'll stop grandstanding and get to the point. Even if the app gets only ~5,000 daily users, that will likely double the traffic for about half the websites featured on Tapestry. With doubled traffic they will be able to grow, and with growth they will be able to better compensate authors for their work, and with better compensation, authors will be able to focus more on producing what they love.

So download Tapestry and tell a book lover about it. Trust me, they'll appreciate it. -pj